requirements.txt
mysql-connector-python==8.0.26
pandas==1.3.5
re
datetime
pandas